Number and text input stories (UIE-199).
