+++ src/stories/src/components/NumberInput.stories.tsx
@@ -1,11 +1,11 @@
-// import { useArgs } from '@storybook/preview-api';
+import { useArgs } from '@storybook/preview-api';
 import type { Meta, StoryFn } from '@storybook/react';
 import React from 'react';
-import { TextInput } from 'src/components/input';
+import { NumberInput, TextInput } from 'src/components/input';
 
-const meta: Meta<typeof TextInput> = {
-  title: 'src/Components/TextInput',
-  component: TextInput,
+const meta: Meta<typeof NumberInput> = {
+  title: 'src/Components/NumberInput',
+  component: NumberInput,
   parameters: {
     layout: 'centered',
   },
@@ -15,13 +15,25 @@ const meta: Meta<typeof TextInput> = {
       control: 'text',
       description: 'Placeholder text for when the input is empty',
     },
-    value: {
+    tooltip: {
       control: 'text',
+      description:
+        'Tooltip to display on hover (right side only). It will also be used as the aria label if aria-label is not specified.',
+    },
+    value: {
+      control: 'number',
       description: 'Value for the field',
     },
     disabled: {
       control: 'boolean',
       description: 'Should the input be disabled?',
+      table: {
+        defaultValue: { summary: 'false' },
+      },
+    },
+    onlyInteger: {
+      control: 'boolean',
+      description: 'Should the input be limited to integers?',
       table: {
         defaultValue: { summary: 'false' },
       },
@@ -57,8 +69,10 @@ const meta: Meta<typeof TextInput> = {
     },
   },
   args: {
-    placeholder: 'Enter a value',
-    'aria-label': 'Non-visible label for accessibility',
+    tooltip: 'This is a number input',
+    min: 0,
+    max: 10,
+    isClearable: false,
   },
 };
 
@@ -70,7 +84,7 @@ export const Example: StoryFn<typeof TextInput> = (props) => {
     updateArgs({ value });
   };
   return (
-    <TextInput
+    <NumberInput
       value={props.value}
       placeholder={props.placeholder}
       disabled={props.disabled}
@@ -79,6 +93,11 @@ export const Example: StoryFn<typeof TextInput> = (props) => {
       id={props.id}
       autoFocus={props.autoFocus}
       style={props.style}
+      min={props.min}
+      max={props.max}
+      isClearable={props.isClearable}
+      onlyInteger={props.onlyInteger}
+      tooltip={props.tooltip}
     />
   );
 };
